+++ servicio-donaciones/src/main/java/com/tp/donatrack/services/MatchmakingService.java
@@ -5,12 +5,14 @@ import com.tp.donatrack.domain.asignacion.ServicioMatchmaking;
 import com.tp.donatrack.domain.donacion.DonacionSegmentada;
 import com.tp.donatrack.domain.donacion.EstadoDonacionSegmentada;
 import com.tp.donatrack.domain.entidad.EntidadBeneficiaria;
+
 import com.tp.donatrack.dtos.EntidadRankingDTO;
 import com.tp.donatrack.dtos.ResultadoMatchmakingDTO;
+
 import com.tp.donatrack.repositories.DonacionRepository;
 import com.tp.donatrack.repositories.EntidadBeneficiariaRepository;
-import org.springframework.stereotype.Service;
 
+import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -39,7 +41,7 @@ public class MatchmakingService {
     /**
      * Ejecuta el matchmaking para una donación segmentada específica.
      */
-    public ResultadoMatchmakingDTO obtenerRanking(Integer donacionSegmentadaId) {
+    public ResultadoMatchmakingDTO obtenerRanking(Long donacionSegmentadaId) {
         DonacionSegmentada segmentada = donacionRepository.findSegmentadaById(donacionSegmentadaId);
         if (segmentada == null) {
             throw new IllegalArgumentException("No se encontró la donación segmentada con ID: " + donacionSegmentadaId);
@@ -80,7 +82,7 @@ public class MatchmakingService {
     /**
      * Asigna una donación segmentada a la entidad beneficiaria elegida del ranking.
      */
-    public void asignarDonacion(Integer donacionSegmentadaId, Long entidadBeneficiariaId) {
+    public void asignarDonacion(Long donacionSegmentadaId, Long entidadBeneficiariaId) {
         DonacionSegmentada segmentada = donacionRepository.findSegmentadaById(donacionSegmentadaId);
         if (segmentada == null) {
             throw new IllegalArgumentException("No se encontró la donación segmentada con ID: " + donacionSegmentadaId);
@@ -100,7 +102,7 @@ public class MatchmakingService {
 
     private ResultadoMatchmakingDTO mapearResultado(DonacionSegmentada segmentada, ResultadoMatchmaking resultado) {
         return ResultadoMatchmakingDTO.builder()
-                .donacionSegmentadaId(segmentada.getId())
+                .donacionSegmentadaId(Math.toIntExact(segmentada.getId()))
                 .subCategoria(segmentada.getSubCategoria() != null ? segmentada.getSubCategoria().getDescripcion()
                         : "Sin Categoría")
                 .cantidadBienes(segmentada.getCantidad())
